forward

+++ v2/coef.cpp
@@ -2,8 +2,8 @@
 #include <cmath>
 #include "coef.h"
 using namespace std;
-double u(double x) {
-	double u1, u2;
+long double u(long double x) {
+	long double u1, u2;
 	u1 = 1.0;
 	u2 = 0.25 * u1;
 	if (x < 0.0) {
@@ -13,12 +13,12 @@ double u(double x) {
 	}
 }
 
-double kappa(double x, double y) {
-	double kappa0 = 1;
+long double kappa(long double x, long double y) {
+	long double kappa0 = 1;
 	return kappa0 * exp(y);
 }
 
-double Qinj(double x, double dx, double y, double dy) {
+long double Qinj(long double x, long double dx, long double y, long double dy) {
 //	cout<<x<<dx<<y<<dy;
 	if ((x < 0.1 * dx) && (x > -0.1 * dx) && (y < 0.5 * dy) && (y > -0.5 * dy)) {
 //		cout<<"Hello!";
