+++ v2/progon.cpp
@@ -1,6 +1,6 @@
 #include "progon.h"
 #include "data_storage.h"
-void matrix_solver(double *A, double *C, double *B, int N, double *F, double *x, double *alpha, double *beta) {
+void matrix_solver(long double *A, long double *C, long double *B, int N, long double *F, long double *x, long double *alpha, long double *beta) {
     alpha[1] = -(B[1] / C[1]);
     beta[1] = (F[1] / C[1]);
     for (int i = 2; i <= N; ++i) {
